Classification of single nucleotide polymorphism (SNP) genotype data
This is
a simulated dataset of single nucleotide polymorphism (SNP) genotype
data
containing 29623 SNPs (total features). Amongst all SNPs are 15 causal
ones which means they and neighboring ones discriminate between case and

controls while remainder are noise.

In the training are 4000 cases and 4000 controls. The task is to predict

the labels of 2000 test individuals and obtained an accuracy of 67%. I
have used fisher score for feature selection and liblinear for
classification